+++ src/main/java/racingcar/Main.java
@@ -1,9 +1,11 @@
 package racingcar;
 
+import racingcar.racing.Racing;
+
 public class Main {
     public static void main(String[] args) {
         Racing racingGame = new Racing();
-        
+
         racingGame.init();
     }
 }
