+++ src/main/java/racingcar/car/CarAction.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 public enum CarAction {
     STAY, FORWARD
