+++ src/test/java/racingcar/CarTest.java
@@ -1,6 +1,9 @@
 package racingcar;
 
 import org.junit.jupiter.api.Test;
+import racingcar.car.Car;
+import racingcar.car.CarAction;
+import racingcar.car.CarDTO;
 
 import static org.assertj.core.api.Assertions.assertThat;
