+++ src/test/java/racingcar/ValidatorTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.racing.Validator;
 
 import java.util.Arrays;
